🔧 Fix model dtype conversion in GRPOTrainer's input preparation

+++ trl/trainer/grpo_trainer.py
@@ -1378,7 +1378,7 @@ class GRPOTrainer(Trainer):
             padding_side="left",
             add_special_tokens=False,
             **kwargs,
-        ).to(self.model.dtype)
+        )
         prompt_inputs = super()._prepare_inputs(prompt_inputs)
         prompt_ids, prompt_mask = prompt_inputs["input_ids"], prompt_inputs["attention_mask"]
